backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "pyjwt[crypto]>=2.12.1",
    "uvicorn[standard]>=0.42.0",
]
